move claims.rs to lib.rs
